feat: add CompanyName schema (#45)
[Ticket](https://trello.com/c/fGNpg2Gr/5277-change-label-for-employercredential)
<!---
Link to the Trello ticket for this work. There should _usually_ be a 1:1
relationship between a ticket and a PR, but that won't always be the
case, so you may link the same ticket to multiple PRs, or add additional
ticket links to this PR. If there is no ticket, you should probably
create one and use the "added after sprint planning" label,
--->

## Summary
<!---
1-2 sentences summarizing the changes included in this PR
--->
Adds CompanyName schema and uses it for EmployerCredential (because we
don't want to display the label from the LegalName schema)
## Changes
<!---
List all non-trivial changes included in this PR. Bullet points are fine
or the individual commits that make up this PR.
--->
- added CompanyName schema
- changed EmployerCredential to use CompanyName schema
## Testing
<!---
How did you test your changes? How might someone else test them?
--->
- tested locally with schema resolver, wallet
## Checklist

- [x] I have performed a self-review of my code
~- [ ] I have commented my code, particularly in hard-to-understand
areas _to a borderline excessive amount_~
~- [ ] I have made any relevant corresponding changes to the
documentation including the project readme.~
- [x] I have run and tested the changes locally
~- [ ] If it is a core feature, I have added an appropriate amount of
unit tests.~
~- [ ] Any dependent changes have been merged and published in upstream
projects~